requirements.txt
Django==2.1
django-jet==1.0.8
django-mdeditor==0.1.14
gunicorn==19.9.0
Markdown==3.1.1
Pillow==6.1.0
pytz==2019.1
sqlparse==0.3.0
